+++ jeap-deploymentlog-domain/src/main/java/ch/admin/bit/jeap/deploymentlog/domain/ComponentVersion.java
@@ -32,6 +32,7 @@ public class ComponentVersion {
     private String commitRef;
 
     @NonNull
+    @Column(nullable = false)
     private ZonedDateTime committedAt;
 
     private boolean publishedVersion;