+++ jeap-deploymentlog-domain/src/main/java/ch/admin/bit/jeap/deploymentlog/domain/FlowStageProperties.java
@@ -1,0 +1,16 @@
+package ch.admin.bit.jeap.deploymentlog.domain;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.stereotype.Component;
+
+@Getter
+@Setter
+@Component
+@ConfigurationProperties(prefix = "jeap.deploymentlog.flow", ignoreUnknownFields = false)
+public class FlowStageProperties {
+
+    private String startEnvironment;
+    private String defaultFinalDeploymentEnvironment;
+}