+++ jeap-deploymentlog-domain/src/main/java/ch/admin/bit/jeap/deploymentlog/domain/exception/InvalidFlowStageConfigurationException.java
@@ -1,0 +1,8 @@
+package ch.admin.bit.jeap.deploymentlog.domain.exception;
+
+public class InvalidFlowStageConfigurationException extends RuntimeException {
+
+    public InvalidFlowStageConfigurationException(String message) {
+        super(message);
+    }
+}